+++ data/types/bolus/combination/combination_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/tidepool-org/platform/data/types/bolus"
 	"github.com/tidepool-org/platform/data/types/bolus/combination"
 	dataTypesBolusCombinationTest "github.com/tidepool-org/platform/data/types/bolus/combination/test"
+	"github.com/tidepool-org/platform/data/types/bolus/prescriptor"
 	dataTypesTest "github.com/tidepool-org/platform/data/types/test"
 	errorsTest "github.com/tidepool-org/platform/errors/test"
 	"github.com/tidepool-org/platform/pointer"
@@ -1149,6 +1150,9 @@ var _ = Describe("Combination", func() {
 						datum := dataTypesBolusCombinationTest.NewCombination()
 						mutator(datum)
 						expectedDatum := dataTypesBolusCombinationTest.CloneCombination(datum)
+						if *datum.Prescriptor.Prescriptor == prescriptor.ManualPrescriptor {
+							expectedDatum.InsulinOnBoard = nil
+						}
 						normalizer := dataNormalizer.New()
 						Expect(normalizer).ToNot(BeNil())
 						datum.Normalize(normalizer.WithOrigin(origin))